[PL] Unify GenericNeumann and GenericNonuniformN.
These are almost the same and the difference is in the
handling of the shape functions.
